Change core package to io.github.bivashy.wttj
