+++ .gitignore
@@ -31,6 +31,9 @@ buildNumber.properties
 .idea/misc.xml
 .idea/vcs.xml
 
+# Ebean generated files
+.idea/ebean-plugin.xml
+
 # AWS User-specific
 .idea/**/aws.xml
