--- core/src/main/java/io/github/bivashy/command/api/Actor.java
@@ -1,9 +1,0 @@
-package io.github.bivashy.command.api;
-
-public interface Actor {
-
-    long getId();
-
-    void reply(String message);
-
-}
