--- core/src/main/java/io/github/bivashy/command/api/ExtendedCommand.java
@@ -1,7 +1,0 @@
-package io.github.bivashy.command.api;
-
-public interface ExtendedCommand<T extends Actor> {
-
-    T actor();
-
-}
