+++ core/src/main/java/io/github/bivashy/wttj/api/command/Actor.java
@@ -1,0 +1,9 @@
+package io.github.bivashy.wttj.api.command;
+
+public interface Actor {
+
+    long getId();
+
+    void reply(String message);
+
+}
