+++ core/src/main/java/io/github/bivashy/wttj/api/command/ExtendedCommand.java
@@ -1,0 +1,7 @@
+package io.github.bivashy.wttj.api.command;
+
+public interface ExtendedCommand<T extends Actor> {
+
+    T actor();
+
+}
