--- database/src/main/java/io/github/bivashy/App.java
@@ -1,13 +1,0 @@
-package io.github.bivashy;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
