+++ database/src/main/java/io/github/bivashy/wttj/App.java
@@ -1,0 +1,13 @@
+package io.github.bivashy.wttj;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
