+++ database/src/test/java/io/github/bivashy/wttj/AppTest.java
@@ -1,0 +1,20 @@
+package io.github.bivashy.wttj;
+
+import org.junit.jupiter.api.*;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+/**
+ * Unit test for simple App.
+ */
+public class AppTest {
+
+    /**
+     * Rigorous Test :-)
+     */
+    @Test
+    public void shouldAnswerWithTrue() {
+        assertTrue(true);
+    }
+
+}
