--- telegram-bot/src/main/java/io/github/bivashy/telegram/bot/properties/TelegramBotProperties.java
@@ -1,25 +1,0 @@
-package io.github.bivashy.telegram.bot.properties;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.util.StringUtils;
-
-@Configuration
-@ConfigurationProperties("telegram")
-public class TelegramBotProperties {
-
-    private String token;
-
-    public String getToken() {
-        return token;
-    }
-
-    public void setToken(String token) {
-        this.token = token;
-    }
-
-    public boolean hasToken() {
-        return StringUtils.hasText(token);
-    }
-
-}
