+++ telegram-bot/src/main/java/io/github/bivashy/wttj/telegram/bot/TelegramBotApplication.java
@@ -1,0 +1,13 @@
+package io.github.bivashy.wttj.telegram.bot;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TelegramBotApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(TelegramBotApplication.class, args);
+    }
+
+}
