--- whatsapp-telegram-transcoder/src/test/java/io/github/bivashy/AppTest.java
@@ -1,20 +1,0 @@
-package io.github.bivashy;
-
-import static org.junit.Assert.assertTrue;
-
-import org.junit.Test;
-
-/**
- * Unit test for simple App.
- */
-public class AppTest 
-{
-    /**
-     * Rigorous Test :-)
-     */
-    @Test
-    public void shouldAnswerWithTrue()
-    {
-        assertTrue( true );
-    }
-}
